+++ R/Result.R
@@ -1,11 +1,11 @@
 #' @include Connection.R
 NULL
 
-setClass("LoggingDBIResult")
+setClass("dblogResult")
 
 make_result_class <- function(base_class) {
 
-  template_name <- "LoggingDBIResult"
+  template_name <- "dblogResult"
   class_name <- paste0(template_name, "-", base_class)
   all_base_classes <- c(template_name, base_class)
 
@@ -29,7 +29,7 @@ make_result_class <- function(base_class) {
   setMethod(
     "show", class_name,
     function(object) {
-      cat("<LoggingDBIResult>\n")
+      cat("<dblogResult>\n")
       show(object@res)
     })