+++ NEWS.md
@@ -28,7 +28,7 @@
 
 Initial prototype. Exported functions:
 
-- `LoggingDBI()` wraps an arbitrary DBI driver with a custom logger.
+- `dblog()` wraps an arbitrary DBI driver with a custom logger.
 - `get_default_logger()` is the default console logger. 
 - `make_console_logger()` creates a new console logger (for internal use only).
 - `make_collect_logger()` creates a new logger that collects logging output and emits all at once.